requirements.txt
django == 3.0.4
django-widget-tweaks == 1.4.5
Celery == 4.4.1
requests == 2.23.0
python-decouple == 3.3
pyyaml == 5.3
ua-parser == 0.10.0
user-agents == 2.1
django-user-agents == 0.4.0


